lemur_compiler: rename the two compilers
This commit changes the names of two Lemur compilers, and revises
them. Also, it moves one additional util file to util/.
